Move CustomVocabulary Model to Shared
